Added Store is open / close button with variable to be passed to other files
