+++ app/admin/page.tsx
@@ -9,7 +9,7 @@ import AboutUs from "./aboutUs/AboutUs";
 import OurWork from "./ourWork/OurWork";
 import Calendar from "../Calendar/admin-calendar/AdminCalendar";
 import Donations from "./donations/Donations";
-import Store from "./store/Store";
+import Store from "./storeAdmin/Store";
 import News from "./news/News";
 import WebsiteSettings from "./WebsiteSettings";
 import Documents from "./documents/Documents";
