+++ app/admin/storeAdmin/store.css
@@ -1,0 +1,52 @@
+/* store.css */
+.store-toggle-btn {
+    padding: 10px 15px;
+    font-size: 16px;
+    cursor: pointer;
+    background-color: black; /* Default background color */
+    color: white;
+    border: none;
+    border-radius: 4px;
+    transition: background-color 0.3s;
+}
+
+/* Hover effect */
+.store-toggle-btn:hover {
+    background-color: #333; /* Darker shade on hover */
+}
+
+/* Store open and closed states */
+.store-toggle-btn.open {
+    background-color: #000000; /* Green when store is open */
+}
+
+.store-toggle-btn.closed {
+    background-color: #000000; /* Red when store is closed */
+}
+
+/* Hover effects for the dynamic states */
+.store-toggle-btn.open:hover {
+    background-color: #45a049; /* Darker green when store is open */
+}
+
+.store-toggle-btn.closed:hover {
+    background-color: #d32f2f; /* Darker red when store is closed */
+}
+
+.collapsible-content {
+    margin-top: 10px;
+    padding-left: 20px;
+    transition: max-height 0.3s ease-out;
+}
+
+.collapsible-content.collapsed {
+    display: none;
+}
+
+.collapsible-content.expanded {
+    display: block;
+}
+
+.dropdown-arrow {
+    font-size: 18px;
+}
